requirements.txt
flask==2.3.3
supabase==1.2.0
python-dotenv==1.0.0
pydantic==2.4.2
email-validator==2.1.0
flask-limiter==3.5.0
flask-cors==4.0.0
PyJWT==2.8.0
logging==0.4.9.6
gunicorn==21.2.0
pytest==7.4.2
black==23.9.1
isort==5.12.0
flake8==6.1.0